+++ blaze-core/src/test/scala/org/http4s/blazecore/websocket/WSTestHead.scala
@@ -1,7 +1,6 @@
 package org.http4s.blazecore.websocket
 
 import cats.effect.{ContextShift, IO, Timer}
-import fs2.Stream
 import fs2.concurrent.Queue
 import org.http4s.blaze.pipeline.HeadStage
 import org.http4s.websocket.WebsocketBits.WebSocketFrame
@@ -61,13 +60,10 @@ sealed abstract class WSTestHead(
       }
 
   def pollBatch(batchSize: Int, timeoutSeconds: Long): IO[List[WebSocketFrame]] =
-    IO.race(
-        timer.sleep(timeoutSeconds.seconds),
-        Stream(batchSize).through(outQueue.dequeueBatch).compile.toList)
-      .map {
-        case Left(_) => Nil
-        case Right(s) => s
-      }
+    outQueue
+      .dequeueChunk1(batchSize)
+      .map(_.toList)
+      .timeoutTo(timeoutSeconds.seconds, IO.pure(Nil))
 
   override def name: String = "WS test stage"
 }